Refactor handleSubmit
